Changed floats to doubles in MoveBy signature.

+++ bebop_driver/include/bebop_driver/bebop.h
@@ -170,7 +170,7 @@ public:
 
   // -1..1
   void Move(const double& roll, const double& pitch, const double& gaz_speed, const double& yaw_speed);
-  void MoveBy(const float& dX, const float& dY, const float& dZ, const float& dPsi);
+  void MoveBy(const double& dX, const double& dY, const double& dZ, const double& dPsi);
   void MoveCamera(const double& tilt, const double& pan);
 
   void TakeSnapshot();
